+++ bin-proto/tests/tag.rs
@@ -22,9 +22,19 @@ pub struct OptionalWriteValue {
     pub data: Vec<u32>,
 }
 
+// Deprecated
+#[derive(BitDecode, BitEncode, Debug, PartialEq, Eq)]
+pub struct WithElementsLengthAutoSelf {
+    #[bin_proto(write_value = self.data.len() as u32)]
+    pub count: u32,
+    pub foo: bool,
+    #[bin_proto(tag = count as usize)]
+    pub data: Vec<u32>,
+}
+
 #[derive(BitDecode, BitEncode, Debug, PartialEq, Eq)]
 pub struct WithElementsLengthAuto {
-    #[bin_proto(write_value = self.data.len() as u32)]
+    #[bin_proto(write_value = data.len() as u32)]
     pub count: u32,
     pub foo: bool,
     #[bin_proto(tag = count as usize)]
@@ -44,15 +54,29 @@ pub enum WithElementsLengthAutoEnum {
     },
 }
 
+// Deprecated
 #[derive(BitDecode, BitEncode, Debug, PartialEq, Eq)]
-pub struct Prepended {
+pub struct PrependedSelf {
     #[bin_proto(tag_type = u32, tag_value = self.data.len() as u32)]
     pub data: Vec<u32>,
 }
 
 #[derive(BitDecode, BitEncode, Debug, PartialEq, Eq)]
-pub struct PrependedBits {
+pub struct Prepended {
+    #[bin_proto(tag_type = u32, tag_value = data.len() as u32)]
+    pub data: Vec<u32>,
+}
+
+// Deprecated
+#[derive(BitDecode, BitEncode, Debug, PartialEq, Eq)]
+pub struct PrependedBitsSelf {
     #[bin_proto(tag_type = u32, tag_value = self.data.len() as u32, tag_bits = 3)]
+    pub data: Vec<u32>,
+}
+
+#[derive(BitDecode, BitEncode, Debug, PartialEq, Eq)]
+pub struct PrependedBits {
+    #[bin_proto(tag_type = u32, tag_value = data.len() as u32, tag_bits = 3)]
     pub data: Vec<u32>,
 }